cleanup empty notice + move tag assign options
